WIP: Move code to common lib
