--- projects/app/src/app/common/controllers/attributes/index.ts
@@ -1,2 +1,0 @@
-export * from './attributes';
-export * from './types';
