--- projects/app/src/app/common/controllers/attributes/types.ts
@@ -1,3 +1,0 @@
-export type HTMLAttributeValue = string | number | boolean;
-
-export type HTMLAttributes = { [attributeName: string]: HTMLAttributeValue };
