+++ projects/common/src/constants/day-duration.ts
