+++ projects/common/src/public-api.ts
@@ -1,5 +1,9 @@
 /*
  * Public API Surface of common
  */
+export * from './constants/public-api';
+export * from './controllers/public-api';
+export * from './rxjs/public-api';
+export * from './store/public-api';
 export * from './types/public-api';
 export * from './utils/public-api';
