+++ projects/common/src/rxjs/first-false.ts
