+++ projects/common/src/rxjs/first-having.ts
