+++ projects/common/src/rxjs/first-non-null.ts
