+++ projects/common/src/rxjs/first-truthy.ts
