+++ projects/common/src/store/collection/collection.ts
@@ -6,12 +6,8 @@ import {
   UnixTimestamp,
 } from '@common/types';
 
-export type SortingFn<T = any> = (a: T, b: T) => -1 | 0 | 1;
-
-export type CategorizedItems<T extends Record<string, any>> = {
-  category: string;
-  items: T[];
-};
+import { CategorizedItems } from './types';
+import { compareByAscendingKey } from './compare';
 
 export function shouldFetchCollection<T extends Record<string, any>>(
   items: T[],
@@ -74,39 +70,10 @@ export function getItemByName<T extends Record<string, any> & { name: string }>(
   return item ?? null;
 }
 
-export function filterItemsByQuery(items: string[], _query: string): string[] {
-  const query = _query.toLowerCase();
-  return items.filter((item) => item.toLowerCase().includes(query));
-}
-
-export function filterItemsByName<
-  T extends Record<string, any> & { name: string }
->(items: T[], name: string): T[] {
-  const query = name.toLowerCase();
-  return items.filter((item) => item.name.toLowerCase().includes(query));
-}
-
 export function sortItemsByName<T extends { name: string }>(items: T[]): T[] {
   const newItems = [...items];
   newItems.sort(compareByAscendingKey('name'));
   return newItems;
-}
-
-export function compareByAscendingKey<T extends Record<string, any>>(
-  key: string
-): SortingFn<T> {
-  const getter = (item: T): T[typeof key] => item[key];
-
-  return (a: T, b: T): -1 | 0 | 1 => {
-    const aValue = getter(a);
-    const bValue = getter(b);
-
-    if (aValue === bValue) {
-      return 0;
-    }
-
-    return aValue < bValue ? -1 : 1;
-  };
 }
 
 export function groupItemsByCategory<T extends { category?: string }>(
